// src/views/expense/voucher/cash.voucher.type.ts
export interface CashVoucherType {
  id: number;
  date: any;
  amount: number | string;
  reason: string;
  operator: string;
  person: any;
  personId: any;
  category: any;
  categoryId: any;
  department: any;
  departmentId: any;
  createdAt: any;
  updatedAt: any;
  status: string;
  usedAmount: number | string;
  number: string;
  personType: string;
}